test:[#134] Service Test를 두 개로 분리
